+++ winnow/hash/__init__.py
@@ -19,6 +19,13 @@ from winnow.hash.cache import HashCache, open_hash_cache
 from winnow.hash.cache_entry import CacheEntry
 from winnow.hash.cache_key import CacheKey
 from winnow.hash.cache_stats import CacheStats
+from winnow.hash.content import (
+    CONTENT_ALGORITHMS,
+    DEFAULT_CHUNK_SIZE,
+    ContentHash,
+    ContentHasher,
+    ContentHasherProtocol,
+)
 from winnow.hash.digest import hamming_distance, parse_digest
 from winnow.hash.hash_failure import HashFailure
 from winnow.hash.hashed_media import HashedMedia
@@ -49,4 +56,9 @@ __all__ = [
     "HashFailure",
     "HashedMedia",
     "hash_media_files",
+    "CONTENT_ALGORITHMS",
+    "DEFAULT_CHUNK_SIZE",
+    "ContentHash",
+    "ContentHasher",
+    "ContentHasherProtocol",
 ]